Biopax example data

A dataset containing Porphyrin metabolism pathway in Biopax Level 3 and parsed with readBiopax.

Examples

data(ex_biopax)
ex_biopax
#> Summary of the biopax object:
#>             Length Class                             Mode     
#> namespaces  5      SimplifiedXMLNamespaceDefinitions character
#> ns_rdf      1      -none-                            character
#> ns_owl      1      -none-                            character
#> ns_bp       1      -none-                            character
#> file        1      -none-                            character
#> biopaxlevel 1      -none-                            numeric  
#> dt          6      biopax_df                         list     
#> 
#> Internal data:
#> $namespaces
#>                                                rdf 
#>      "http://www.w3.org/1999/02/22-rdf-syntax-ns#" 
#>                                                 bp 
#> "http://www.biopax.org/release/biopax-level3.owl#" 
#>                                                owl 
#>                   "http://www.w3.org/2002/07/owl#" 
#>                                               rdfs 
#>            "http://www.w3.org/2000/01/rdf-schema#" 
#>                                                xsd 
#>                "http://www.w3.org/2001/XMLSchema#" 
#> attr(,"class")
#> [1] "SimplifiedXMLNamespaceDefinitions" "XMLNamespaceDefinitions"          
#> 
#> $ns_rdf
#> [1] "rdf"
#> 
#> $ns_owl
#> [1] "owl"
#> 
#> $ns_bp
#> [1] "bp"
#> 
#> $file
#> [1] "porph.bio3.owl"
#> 
#> $biopaxlevel
#> [1] 3
#> 
#> $dt
#>                     class                   id  property property_attr
#>    1:           BioSource           BioSource1      name  rdf:datatype
#>    2:           BioSource           BioSource1      xref  rdf:resource
#>    3: BiochemicalReaction BiochemicalReaction1   comment  rdf:datatype
#>    4: BiochemicalReaction BiochemicalReaction1   comment  rdf:datatype
#>    5: BiochemicalReaction BiochemicalReaction1   comment  rdf:datatype
#>   ---                                                                 
#> 3005:     UnificationXref    UnificationXref98        id  rdf:datatype
#> 3006:     UnificationXref    UnificationXref98 idVersion  rdf:datatype
#> 3007:     UnificationXref    UnificationXref99   comment  rdf:datatype
#> 3008:     UnificationXref    UnificationXref99        db  rdf:datatype
#> 3009:     UnificationXref    UnificationXref99        id  rdf:datatype
#>                           property_attr_value
#>    1: http://www.w3.org/2001/XMLSchema#string
#>    2:                      #UnificationXref17
#>    3: http://www.w3.org/2001/XMLSchema#string
#>    4: http://www.w3.org/2001/XMLSchema#string
#>    5: http://www.w3.org/2001/XMLSchema#string
#>   ---                                        
#> 3005: http://www.w3.org/2001/XMLSchema#string
#> 3006: http://www.w3.org/2001/XMLSchema#string
#> 3007: http://www.w3.org/2001/XMLSchema#string
#> 3008: http://www.w3.org/2001/XMLSchema#string
#> 3009: http://www.w3.org/2001/XMLSchema#string
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  property_value
#>    1:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              Homo sapiens
#>    2:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          
#>    3: The committed step for porphyrin synthesis is the formation of 5-aminolevulinate (ALA) by condensation of glycine (from the general amino acid pool) and succinyl-CoA (from the TCA cycle), in the mitochondrial matrix. The reaction is catalyzed by two different ALA synthases, one expressed ubiquitously (ALAS1) and the other only expressed in erythroid precursors (ALAS2). Both enzymes are expressed as homodimers and require pyridoxal 5-phosphate as a cofactor.<br>No disease-causing mutations of ALAS1 have been recognised in humans. Mutations in ALAS2 cause X-linked sideroblastic anaemia (XLSA), characterised by a microcytic hypochromic anaemia.
#>    4:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  Authored: Jassal, B, D'Eustachio, P, 2007-01-24 10:19:49
#>    5:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   Reviewed: Sassa, S, 2007-01-24 10:18:36
#>   ---                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          
#> 3005:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                REACT_9699
#> 3006:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         3
#> 3007:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           Database identifier. Use this URL to connect to the web page of this instance in Reactome: http://www.reactome.org/cgi-bin/eventbrowser?DB=gk_current&ID=113528
#> 3008:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           Reactome Database ID Release 46
#> 3009:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    113528
#> 
#> Dimension of internal data.table:  3009,6 
#> 
#> Summary of parsed internal data.table 
#>     class                id              property         property_attr     
#>  Length:3009        Length:3009        Length:3009        Length:3009       
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>  property_attr_value property_value    
#>  Length:3009         Length:3009       
#>  Class :character    Class :character  
#>  Mode  :character    Mode  :character