Convert a ranked path list to edge ids of a graph

Convert a ranked path list to Edge ids of a graph, where paths can come from a different representation (for example matching path from a reaction network to edges on a metabolic network).

getPathsAsEIDs(paths, graph)

Arguments

paths

The paths extracted by pathRanker.

graph

A annotated igraph object.

Value

A list of edge ids on the provided graph.

See also

Author

Ahmed Mohamed

Examples

  ## Prepare a weighted reaction network.
  ## Conver a metabolic network to a reaction network.
 data(ex_sbml) # bipartite metabolic network of Carbohydrate metabolism.
 rgraph <- makeReactionNetwork(ex_sbml, simplify=TRUE)
#> This graph was created by an old(er) igraph version.
#> ℹ Call `igraph::upgrade_graph()` on it to use with the current igraph version.
#> For now we convert it on the fly...

  ## Assign edge weights based on Affymetrix attributes and microarray dataset.
 # Calculate Pearson's correlation.
  data(ex_microarray)  # Part of ALL dataset.
  rgraph <- assignEdgeWeights(microarray = ex_microarray, graph = rgraph,
    weight.method = "cor", use.attr="miriam.uniprot",
    y=factor(colnames(ex_microarray)), bootstrap = FALSE)
#> 100 genes were present in the microarray, but not represented in the network.
#> 55 genes were couldn't be found in microarray.
#> Assigning edge weights for label ALL1/AF4 
#> Assigning edge weights for label BCR/ABL 
#> Assigning edge weights for label E2A/PBX1 
#> Assigning edge weights for label NEG 

  ## Get ranked paths using probabilistic shortest paths.
 ranked.p <- pathRanker(rgraph, method="prob.shortest.path",
          K=20, minPathSize=6)
#> Extracting the 20 most probable paths for ALL1/AF4
#> Extracting the 20 most probable paths for BCR/ABL
#> Extracting the 20 most probable paths for E2A/PBX1
#> Extracting the 20 most probable paths for NEG

  ## Get the edge ids along paths in the reaction graph.
  path.eids <- getPathsAsEIDs(ranked.p, rgraph)

  ## Get the edge ids along paths in the original metabolic graph.
  path.eids <- getPathsAsEIDs(ranked.p, ex_sbml)