Plots an annotated igraph object in Cytoscape.

plotCytoscape function has been removed because RCytoscape is no longer prensent in Bioconductor. Future plans will use RCy3 for Cytoscape plotting, once RCy3 is supported on MacOS and Windows. plotCytoscapeGML exports the network plot in GML format, that can be later imported into Cytoscape (using "import network from file" option). This fuction is compatible with all Cytoscape versions.

plotCytoscapeGML(graph, file, layout = layout.auto, vertex.size,
  vertex.label, vertex.shape, vertex.color, edge.color)

Arguments

graph

An annotated igraph object.

file

Output GML file name to which the network plot is exported.

layout

Either a graph layout function, or a two-column matrix specifiying vertex coordinates.

vertex.size

Vertex size. If missing, the vertex attribute "size" (

V(g)$size)

) will be used.

vertex.label

Vertex labels. If missing, the vertex attribute "label" (

V(g)$label)

) will be used. If missing, vertices are labeled by their name.

vertex.shape

Vertex shape in one of igraph shapes. If missing, the vertex attribute "shape" (

V(g)$shape)

) will be used. Shapes are converted from igraph convention to Cytoscape convention. "square","rectangle" and "vrectangle" are converted to "RECT", "csquare" and "crectangle" are converted to "ROUND_RECT", all other shapes are considered "ELLIPSE"

vertex.color

A color or a list of colors for vertices. Vetices with multiple colors are not supported. If missing, the vertex attribute "color" (

V(g)$color)

) will be used.

edge.color

A color or a list of colors for edges. If missing, the edge attribute "color" (

E(g)$color)

) will be used.

Value

For plotCytoscapeGML, results are written to file.

See also

Author

Ahmed Mohamed

Examples

 data("ex_sbml")
  rgraph <- makeReactionNetwork(ex_sbml, simplify=TRUE)
 v.layout <- layoutVertexByAttr(rgraph, "compartment")
  v.color <- colorVertexByAttr(rgraph, "compartment")

 # Export network plot to GML file
 plotCytoscapeGML(rgraph, file="example.gml", layout=v.layout,
        vertex.color=v.color, vertex.size=10)
#> Warning: one argument not used by format 'node [
#>  %1$s %4$g
#> %2$s "%5$s"
#> %6$s
#>  ]'
#> Warning: one argument not used by format 'edge [
#>  %1$s %4$g
#> %2$s %5$g
#> %6$s
#>  ]'