Remove uniquitous compounds from a metabolic network

This function removes uniquitous compounds (metabolites connected to numerous reactions) from a metabolic network.These compounds are reaction cofactors and currency compounds, such as ATP, CO2, etc. A path through these metabolites may not be bioloigcally meaningful. The defualt small compound list is derived from Reactome, containing keeg.compound, pubchem.compound, ChEBI and CAS identifiers.

rmSmallCompounds(graph, method = c("remove", "duplicate"),
  small.comp.ls = NPMdefaults("small.comp.ls"))

Arguments

graph

A metabolic network.

method

How to handle small compounds. Either simply delete these vertices "remove" (default), or make a separate vertex for each reaction they participate in "duplicate".

small.comp.ls

A list of small compounds to be used.

Value

A modified graph, with the small compounds removed or duplicated.

See also

Author

Ahmed Mohamed

Examples

    data(ex_sbml)
    ex_sbml
#> IGRAPH 1981d4c DN-- 432 662 -- 
#> + attr: source (g/c), type (g/c), name (v/c), attr (v/x), reactions
#> | (v/l), shape (v/c), color (v/c), stoichiometry (e/n)
#> + edges from 1981d4c (vertex names):
#>  [1] reaction_71850  ->species_113592  reaction_71850  ->species_29728  
#>  [3] reaction_71496  ->species_75970   reaction_71496  ->species_29578  
#>  [5] reaction_2046222->species_2046166 reaction_2046222->species_742343 
#>  [7] reaction_2162229->species_2162232 reaction_2162229->species_2162231
#>  [9] reaction_2162226->species_1678868 reaction_2162226->species_1678743
#> [11] reaction_70403  ->species_70113   reaction_2162227->species_1678868
#> [13] reaction_2162227->species_2162223 reaction_2162225->species_1678743
#> + ... omitted several edges

    sbml.removed <- rmSmallCompounds(ex_sbml, method="remove")
    sbml.removed
#> IGRAPH d6668a7 DN-- 398 574 -- 
#> + attr: source (g/c), type (g/c), name (v/c), attr (v/x), reactions
#> | (v/l), shape (v/c), color (v/c), stoichiometry (e/n)
#> + edges from d6668a7 (vertex names):
#>  [1] reaction_71850  ->species_113592  reaction_71496  ->species_75970  
#>  [3] reaction_71496  ->species_29578   reaction_2046222->species_742343 
#>  [5] reaction_2162229->species_2162232 reaction_2162229->species_2162231
#>  [7] reaction_2162226->species_1678868 reaction_2162226->species_1678743
#>  [9] reaction_2162227->species_1678868 reaction_2162227->species_2162223
#> [11] reaction_2162225->species_1678743 reaction_2162225->species_2162224
#> [13] reaction_1678694->species_2090076 reaction_1678694->species_2090023
#> + ... omitted several edges