Creates a set of sample path p-values for each length given a weighted network

Randomly traverses paths of increasing lengths within a set network to create an empirical pathway distribution for more accurate determination of path significance.

samplePaths(graph, max.path.length, num.samples = 1000,
  num.warmup = 10, verbose = TRUE)

Arguments

graph

A weighted igraph object. Weights must be in edge.weights or weight edge attributes.

max.path.length

The maxmimum path length.

num.samples

The numner of paths to sample

num.warmup

The number of warm up paths to sample.

verbose

Whether to display the progress of the function.

Value

A matrix where each row is a path length and each column is the number of paths sampled.

Details

Can take a bit of time.

See also

Author

Timothy Hancock

Ahmed Mohamed

Examples

  ## Prepare a weighted reaction network.
  ## Conver a metabolic network to a reaction network.
 data(ex_sbml) # bipartite metabolic network of Carbohydrate metabolism.
 rgraph <- makeReactionNetwork(ex_sbml, simplify=TRUE)

  ## Assign edge weights based on Affymetrix attributes and microarray dataset.
 # Calculate Pearson's correlation.
  data(ex_microarray)  # Part of ALL dataset.
  rgraph <- assignEdgeWeights(microarray = ex_microarray, graph = rgraph,
    weight.method = "cor", use.attr="miriam.uniprot",
    y=factor(colnames(ex_microarray)), bootstrap = FALSE)
#> 100 genes were present in the microarray, but not represented in the network.
#> 55 genes were couldn't be found in microarray.
#> Assigning edge weights for label ALL1/AF4 
#> Assigning edge weights for label BCR/ABL 
#> Assigning edge weights for label E2A/PBX1 
#> Assigning edge weights for label NEG 

  ## Get significantly correlated paths using "p-valvue" method.
  ##   First, establish path score distribution by calling "samplePaths"
 pathsample <- samplePaths(rgraph, max.path.length=10,
                        num.samples=100, num.warmup=10)
#> Sampling100 forALL1/AF4
#> Sampling100 forBCR/ABL
#> Sampling100 forE2A/PBX1
#> Sampling100 forNEG

  ##   Get all significant paths with p<0.1
  significant.p <- pathRanker(rgraph, method = "pvalue",
                sampledpaths = pathsample ,alpha=0.1)
#> Start nodes weren't provided. Using entry nodes as start points.
#> Extracting non-random path p<0.1 for ALL1/AF4
#> There are 151 nodes in the neighborhoodTARGET : reaction_71850 1/ 151 - Found a path to reaction_71850
#> TARGET : reaction_71496 2/ 151 - Found a path to reaction_71496
#> TARGET : reaction_2046222 3/ 151 - Node out of scope
#> TARGET : reaction_2162226 4/ 151 - Node out of scope
#> TARGET : reaction_70403 5/ 151 - Found a path to reaction_70403
#> TARGET : reaction_2162227 6/ 151 - Node out of scope
#> TARGET : reaction_2162225 7/ 151 - Node out of scope
#> TARGET : reaction_2090085 8/ 151 - Node out of scope
#> TARGET : reaction_3262512 9/ 151 - Node out of scope
#> TARGET : reaction_372448 10/ 151 - Node out of scope
#> TARGET : reaction_372449 11/ 151 - Node out of scope
#> TARGET : reaction_71495 12/ 151 - Found a path to reaction_71495
#> TARGET : reaction_2076508 13/ 151 - Node out of scope
#> TARGET : reaction_71515 14/ 151 - Node out of scope
#> TARGET : reaction_1638032 15/ 151 - Node out of scope
#> TARGET : reaction_70420 16/ 151 - Found a path to reaction_70420
#> TARGET : reaction_1678716 17/ 151 - Node out of scope
#> TARGET : reaction_1793217 18/ 151 - Node out of scope
#> TARGET : reaction_450092 19/ 151 - Node out of scope
#> TARGET : reaction_450095 20/ 151 - Found a path to reaction_450095
#> TARGET : reaction_1678708 21/ 151 - Node out of scope
#> TARGET : reaction_70427 22/ 151 - Node out of scope
#> TARGET : reaction_450088 23/ 151 - Found a path to reaction_450088
#> TARGET : reaction_3266566 24/ 151 - Node out of scope
#> TARGET : reaction_191101 25/ 151 - Node out of scope
#> TARGET : reaction_198508 26/ 151 - Node out of scope
#> TARGET : reaction_1678660 27/ 151 - Node out of scope
#> TARGET : reaction_1630306 28/ 151 - Node out of scope
#> TARGET : reaction_1630304 29/ 151 - Node out of scope
#> TARGET : reaction_70454 30/ 151 - Found a path to reaction_70454
#> TARGET : reaction_70449 31/ 151 - Node out of scope
#> TARGET : reaction_2046265 32/ 151 - Found a path to reaction_2046265
#> TARGET : reaction_71825 33/ 151 - Node out of scope
#> TARGET : reaction_198513 34/ 151 - Node out of scope
#> TARGET : reaction_71541 35/ 151 - Found a path to reaction_71541
#> TARGET : reaction_70467 36/ 151 - Found a path to reaction_70467
#> TARGET : reaction_1793182 37/ 151 - Found a path to reaction_1793182
#> TARGET : reaction_2022061 38/ 151 - Node out of scope
#> TARGET : reaction_70471 39/ 151 - Found a path to reaction_70471
#> TARGET : reaction_2046285 40/ 151 - Found a path to reaction_2046285
#> TARGET : reaction_70475 41/ 151 - Node out of scope
#> TARGET : reaction_2022052 42/ 151 - Node out of scope
#> TARGET : reaction_70479 43/ 151 - Found a path to reaction_70479
#> TARGET : reaction_2025724 44/ 151 - Found a path to reaction_2025724
#> TARGET : reaction_1678650 45/ 151 - Node out of scope
#> TARGET : reaction_70482 46/ 151 - Node out of scope
#> TARGET : reaction_741449 47/ 151 - Node out of scope
#> TARGET : reaction_70481 48/ 151 - Found a path to reaction_70481
#> TARGET : reaction_2046298 49/ 151 - Node out of scope
#> TARGET : reaction_2090043 50/ 151 - Node out of scope
#> TARGET : reaction_372843 51/ 151 - Node out of scope
#> TARGET : reaction_2022851 52/ 151 - Node out of scope
#> TARGET : reaction_70486 53/ 151 - Found a path to reaction_70486
#> TARGET : reaction_1606789 54/ 151 - Node out of scope
#> TARGET : reaction_1678854 55/ 151 - Node out of scope
#> TARGET : reaction_2105001 56/ 151 - Node out of scope
#> TARGET : reaction_2160915 57/ 151 - Node out of scope
#> TARGET : reaction_177784 58/ 151 - Node out of scope
#> TARGET : reaction_2022860 59/ 151 - Node out of scope
#> TARGET : reaction_71445 60/ 151 - Found a path to reaction_71445
#> TARGET : reaction_2090037 61/ 151 - Node out of scope
#> TARGET : reaction_70494 62/ 151 - Node out of scope
#> TARGET : reaction_2090038 63/ 151 - Node out of scope
#> TARGET : reaction_2022856 64/ 151 - Node out of scope
#> TARGET : reaction_70262 65/ 151 - Found a path to reaction_70262
#> TARGET : reaction_70501 66/ 151 - Node out of scope
#> TARGET : reaction_198440 67/ 151 - Node out of scope
#> TARGET : reaction_372819 68/ 151 - Node out of scope
#> TARGET : reaction_189102 69/ 151 - Node out of scope
#> TARGET : reaction_70241 70/ 151 - Node out of scope
#> TARGET : reaction_2022887 71/ 151 - Node out of scope
#> TARGET : reaction_198458 72/ 151 - Node out of scope
#> TARGET : reaction_2076371 73/ 151 - Node out of scope
#> TARGET : reaction_163750 74/ 151 - Found a path to reaction_163750
#> TARGET : reaction_2065233 75/ 151 - Node out of scope
#> TARGET : reaction_2076611 76/ 151 - Node out of scope
#> TARGET : reaction_163751 77/ 151 - Found a path to reaction_163751
#> TARGET : reaction_2076383 78/ 151 - Node out of scope
#> TARGET : reaction_70290 79/ 151 - Node out of scope
#> TARGET : reaction_70286 80/ 151 - Node out of scope
#> TARGET : reaction_163764 81/ 151 - Found a path to reaction_163764
#> TARGET : reaction_199803 82/ 151 - Found a path to reaction_199803
#> TARGET : reaction_1971487 83/ 151 - Node out of scope
#> TARGET : reaction_163773 84/ 151 - Found a path to reaction_163773
#> TARGET : reaction_70272 85/ 151 - Node out of scope
#> TARGET : reaction_1971483 86/ 151 - Node out of scope
#> TARGET : reaction_70333 87/ 151 - Node out of scope
#> TARGET : reaction_2160884 88/ 151 - Node out of scope
#> TARGET : reaction_1971491 89/ 151 - Node out of scope
#> TARGET : reaction_71654 90/ 151 - Found a path to reaction_71654
#> TARGET : reaction_1889981 91/ 151 - Node out of scope
#> TARGET : reaction_2160892 92/ 151 - Node out of scope
#> TARGET : reaction_71660 93/ 151 - Node out of scope
#> TARGET : reaction_1889978 94/ 151 - Node out of scope
#> TARGET : reaction_189053 95/ 151 - Node out of scope
#> TARGET : reaction_73580 96/ 151 - Node out of scope
#> TARGET : reaction_71670 97/ 151 - Found a path to reaction_71670
#> TARGET : reaction_2024100 98/ 151 - Node out of scope
#> TARGET : reaction_70581 99/ 151 - Node out of scope
#> TARGET : reaction_2076392 100/ 151 - Node out of scope
#> TARGET : reaction_2160874 101/ 151 - Node out of scope
#> TARGET : reaction_163741 102/ 151 - Found a path to reaction_163741
#> TARGET : reaction_70361 103/ 151 - Node out of scope
#> TARGET : reaction_71324 104/ 151 - Node out of scope
#> TARGET : reaction_453337 105/ 151 - Found a path to reaction_453337
#> TARGET : reaction_453338 106/ 151 - Node out of scope
#> TARGET : reaction_70355 107/ 151 - Found a path to reaction_70355
#> TARGET : reaction_70592 108/ 151 - Node out of scope
#> TARGET : reaction_453339 109/ 151 - Found a path to reaction_453339
#> TARGET : reaction_2046175 110/ 151 - Node out of scope
#> TARGET : reaction_453342 111/ 151 - Node out of scope
#> TARGET : reaction_70596 112/ 151 - Node out of scope
#> TARGET : reaction_71574 113/ 151 - Node out of scope
#> TARGET : reaction_174392 114/ 151 - Node out of scope
#> TARGET : reaction_71303 115/ 151 - Node out of scope
#> TARGET : reaction_70349 116/ 151 - Found a path to reaction_70349
#> TARGET : reaction_191108 117/ 151 - Node out of scope
#> TARGET : reaction_71296 118/ 151 - Node out of scope
#> TARGET : reaction_71299 119/ 151 - Node out of scope
#> TARGET : reaction_174389 120/ 151 - Node out of scope
#> TARGET : reaction_71802 121/ 151 - Found a path to reaction_71802
#> TARGET : reaction_191116 122/ 151 - Node out of scope
#> TARGET : reaction_70613 123/ 151 - Node out of scope
#> TARGET : reaction_71306 124/ 151 - Node out of scope
#> TARGET : reaction_70342 125/ 151 - Node out of scope
#> TARGET : reaction_71588 126/ 151 - Found a path to reaction_71588
#> TARGET : reaction_71590 127/ 151 - Found a path to reaction_71590
#> TARGET : reaction_2142859 128/ 151 - Node out of scope
#> TARGET : reaction_71593 129/ 151 - Node out of scope
#> TARGET : reaction_71335 130/ 151 - Node out of scope
#> TARGET : reaction_453346 131/ 151 - Node out of scope
#> TARGET : reaction_71334 132/ 151 - Found a path to reaction_71334
#> TARGET : reaction_70377 133/ 151 - Node out of scope
#> TARGET : reaction_2046180 134/ 151 - Node out of scope
#> TARGET : reaction_71602 135/ 151 - Node out of scope
#> TARGET : reaction_1678742 136/ 151 - Node out of scope
#> TARGET : reaction_70369 137/ 151 - Node out of scope
#> TARGET : reaction_453358 138/ 151 - Found a path to reaction_453358
#> TARGET : reaction_189222 139/ 151 - Node out of scope
#> TARGET : reaction_453356 140/ 151 - Node out of scope
#> TARGET : reaction_1638053 141/ 151 - Node out of scope
#> TARGET : reaction_2022063 142/ 151 - Node out of scope
#> TARGET : reaction_2090079 143/ 151 - Node out of scope
#> TARGET : reaction_1793186 144/ 151 - Node out of scope
#> TARGET : reaction_2076419 145/ 151 - Node out of scope
#> TARGET : reaction_1793207 146/ 151 - Node out of scope
#> TARGET : reaction_1889955 147/ 151 - Node out of scope
#> TARGET : reaction_111215 148/ 151 - Node out of scope
#> TARGET : reaction_71610 149/ 151 - Found a path to reaction_71610
#> TARGET : reaction_2018682 150/ 151 - Node out of scope
#> TARGET : reaction_2018659 151/ 151 - Node out of scope
#> Extracting non-random path p<0.1 for BCR/ABL
#> There are 151 nodes in the neighborhoodTARGET : reaction_71850 1/ 151 - Node out of scope
#> TARGET : reaction_71496 2/ 151 - Node out of scope
#> TARGET : reaction_2046222 3/ 151 - Node out of scope
#> TARGET : reaction_2162226 4/ 151 - Node out of scope
#> TARGET : reaction_70403 5/ 151 - Found a path to reaction_70403
#> TARGET : reaction_2162227 6/ 151 - Node out of scope
#> TARGET : reaction_2162225 7/ 151 - Node out of scope
#> TARGET : reaction_2090085 8/ 151 - Node out of scope
#> TARGET : reaction_3262512 9/ 151 - Node out of scope
#> TARGET : reaction_372448 10/ 151 - Node out of scope
#> TARGET : reaction_372449 11/ 151 - Node out of scope
#> TARGET : reaction_71495 12/ 151 - Node out of scope
#> TARGET : reaction_2076508 13/ 151 - Node out of scope
#> TARGET : reaction_71515 14/ 151 - Node out of scope
#> TARGET : reaction_1638032 15/ 151 - Node out of scope
#> TARGET : reaction_70420 16/ 151 - Found a path to reaction_70420
#> TARGET : reaction_1678716 17/ 151 - Node out of scope
#> TARGET : reaction_1793217 18/ 151 - Node out of scope
#> TARGET : reaction_450092 19/ 151 - Node out of scope
#> TARGET : reaction_450095 20/ 151 - Node out of scope
#> TARGET : reaction_1678708 21/ 151 - Node out of scope
#> TARGET : reaction_70427 22/ 151 - Node out of scope
#> TARGET : reaction_450088 23/ 151 - Node out of scope
#> TARGET : reaction_3266566 24/ 151 - Node out of scope
#> TARGET : reaction_191101 25/ 151 - Node out of scope
#> TARGET : reaction_198508 26/ 151 - Node out of scope
#> TARGET : reaction_1678660 27/ 151 - Node out of scope
#> TARGET : reaction_1630306 28/ 151 - Node out of scope
#> TARGET : reaction_1630304 29/ 151 - Node out of scope
#> TARGET : reaction_70454 30/ 151 - Node out of scope
#> TARGET : reaction_70449 31/ 151 - Node out of scope
#> TARGET : reaction_2046265 32/ 151 - Node out of scope
#> TARGET : reaction_71825 33/ 151 - Node out of scope
#> TARGET : reaction_198513 34/ 151 - Node out of scope
#> TARGET : reaction_71541 35/ 151 - Node out of scope
#> TARGET : reaction_70467 36/ 151 - Node out of scope
#> TARGET : reaction_1793182 37/ 151 - Node out of scope
#> TARGET : reaction_2022061 38/ 151 - Node out of scope
#> TARGET : reaction_70471 39/ 151 - Node out of scope
#> TARGET : reaction_2046285 40/ 151 - Node out of scope
#> TARGET : reaction_70475 41/ 151 - Node out of scope
#> TARGET : reaction_2022052 42/ 151 - Node out of scope
#> TARGET : reaction_70479 43/ 151 - Node out of scope
#> TARGET : reaction_2025724 44/ 151 - Found a path to reaction_2025724
#> TARGET : reaction_1678650 45/ 151 - Node out of scope
#> TARGET : reaction_70482 46/ 151 - Node out of scope
#> TARGET : reaction_741449 47/ 151 - Node out of scope
#> TARGET : reaction_70481 48/ 151 - Node out of scope
#> TARGET : reaction_2046298 49/ 151 - Node out of scope
#> TARGET : reaction_2090043 50/ 151 - Node out of scope
#> TARGET : reaction_372843 51/ 151 - Node out of scope
#> TARGET : reaction_2022851 52/ 151 - Node out of scope
#> TARGET : reaction_70486 53/ 151 - Node out of scope
#> TARGET : reaction_1606789 54/ 151 - Node out of scope
#> TARGET : reaction_1678854 55/ 151 - Node out of scope
#> TARGET : reaction_2105001 56/ 151 - Node out of scope
#> TARGET : reaction_2160915 57/ 151 - Node out of scope
#> TARGET : reaction_177784 58/ 151 - Node out of scope
#> TARGET : reaction_2022860 59/ 151 - Node out of scope
#> TARGET : reaction_71445 60/ 151 - Node out of scope
#> TARGET : reaction_2090037 61/ 151 - Node out of scope
#> TARGET : reaction_70494 62/ 151 - Node out of scope
#> TARGET : reaction_2090038 63/ 151 - Node out of scope
#> TARGET : reaction_2022856 64/ 151 - Node out of scope
#> TARGET : reaction_70262 65/ 151 - Node out of scope
#> TARGET : reaction_70501 66/ 151 - Node out of scope
#> TARGET : reaction_198440 67/ 151 - Node out of scope
#> TARGET : reaction_372819 68/ 151 - Node out of scope
#> TARGET : reaction_189102 69/ 151 - Node out of scope
#> TARGET : reaction_70241 70/ 151 - Node out of scope
#> TARGET : reaction_2022887 71/ 151 - Node out of scope
#> TARGET : reaction_198458 72/ 151 - Node out of scope
#> TARGET : reaction_2076371 73/ 151 - Node out of scope
#> TARGET : reaction_163750 74/ 151 - Node out of scope
#> TARGET : reaction_2065233 75/ 151 - Node out of scope
#> TARGET : reaction_2076611 76/ 151 - Node out of scope
#> TARGET : reaction_163751 77/ 151 - Node out of scope
#> TARGET : reaction_2076383 78/ 151 - Node out of scope
#> TARGET : reaction_70290 79/ 151 - Node out of scope
#> TARGET : reaction_70286 80/ 151 - Node out of scope
#> TARGET : reaction_163764 81/ 151 - Node out of scope
#> TARGET : reaction_199803 82/ 151 - Node out of scope
#> TARGET : reaction_1971487 83/ 151 - Node out of scope
#> TARGET : reaction_163773 84/ 151 - Node out of scope
#> TARGET : reaction_70272 85/ 151 - Node out of scope
#> TARGET : reaction_1971483 86/ 151 - Node out of scope
#> TARGET : reaction_70333 87/ 151 - Node out of scope
#> TARGET : reaction_2160884 88/ 151 - Node out of scope
#> TARGET : reaction_1971491 89/ 151 - Node out of scope
#> TARGET : reaction_71654 90/ 151 - Node out of scope
#> TARGET : reaction_1889981 91/ 151 - Node out of scope
#> TARGET : reaction_2160892 92/ 151 - Node out of scope
#> TARGET : reaction_71660 93/ 151 - Node out of scope
#> TARGET : reaction_1889978 94/ 151 - Node out of scope
#> TARGET : reaction_189053 95/ 151 - Node out of scope
#> TARGET : reaction_73580 96/ 151 - Node out of scope
#> TARGET : reaction_71670 97/ 151 - Found a path to reaction_71670
#> TARGET : reaction_2024100 98/ 151 - Node out of scope
#> TARGET : reaction_70581 99/ 151 - Node out of scope
#> TARGET : reaction_2076392 100/ 151 - Node out of scope
#> TARGET : reaction_2160874 101/ 151 - Node out of scope
#> TARGET : reaction_163741 102/ 151 - Node out of scope
#> TARGET : reaction_70361 103/ 151 - Node out of scope
#> TARGET : reaction_71324 104/ 151 - Node out of scope
#> TARGET : reaction_453337 105/ 151 - Node out of scope
#> TARGET : reaction_453338 106/ 151 - Node out of scope
#> TARGET : reaction_70355 107/ 151 - Node out of scope
#> TARGET : reaction_70592 108/ 151 - Node out of scope
#> TARGET : reaction_453339 109/ 151 - Node out of scope
#> TARGET : reaction_2046175 110/ 151 - Node out of scope
#> TARGET : reaction_453342 111/ 151 - Node out of scope
#> TARGET : reaction_70596 112/ 151 - Node out of scope
#> TARGET : reaction_71574 113/ 151 - Node out of scope
#> TARGET : reaction_174392 114/ 151 - Node out of scope
#> TARGET : reaction_71303 115/ 151 - Node out of scope
#> TARGET : reaction_70349 116/ 151 - Node out of scope
#> TARGET : reaction_191108 117/ 151 - Node out of scope
#> TARGET : reaction_71296 118/ 151 - Node out of scope
#> TARGET : reaction_71299 119/ 151 - Node out of scope
#> TARGET : reaction_174389 120/ 151 - Node out of scope
#> TARGET : reaction_71802 121/ 151 - Node out of scope
#> TARGET : reaction_191116 122/ 151 - Node out of scope
#> TARGET : reaction_70613 123/ 151 - Node out of scope
#> TARGET : reaction_71306 124/ 151 - Node out of scope
#> TARGET : reaction_70342 125/ 151 - Node out of scope
#> TARGET : reaction_71588 126/ 151 - Node out of scope
#> TARGET : reaction_71590 127/ 151 - Node out of scope
#> TARGET : reaction_2142859 128/ 151 - Node out of scope
#> TARGET : reaction_71593 129/ 151 - Node out of scope
#> TARGET : reaction_71335 130/ 151 - Node out of scope
#> TARGET : reaction_453346 131/ 151 - Node out of scope
#> TARGET : reaction_71334 132/ 151 - Node out of scope
#> TARGET : reaction_70377 133/ 151 - Node out of scope
#> TARGET : reaction_2046180 134/ 151 - Node out of scope
#> TARGET : reaction_71602 135/ 151 - Node out of scope
#> TARGET : reaction_1678742 136/ 151 - Node out of scope
#> TARGET : reaction_70369 137/ 151 - Node out of scope
#> TARGET : reaction_453358 138/ 151 - Node out of scope
#> TARGET : reaction_189222 139/ 151 - Node out of scope
#> TARGET : reaction_453356 140/ 151 - Node out of scope
#> TARGET : reaction_1638053 141/ 151 - Node out of scope
#> TARGET : reaction_2022063 142/ 151 - Node out of scope
#> TARGET : reaction_2090079 143/ 151 - Node out of scope
#> TARGET : reaction_1793186 144/ 151 - Node out of scope
#> TARGET : reaction_2076419 145/ 151 - Node out of scope
#> TARGET : reaction_1793207 146/ 151 - Node out of scope
#> TARGET : reaction_1889955 147/ 151 - Node out of scope
#> TARGET : reaction_111215 148/ 151 - Node out of scope
#> TARGET : reaction_71610 149/ 151 - Node out of scope
#> TARGET : reaction_2018682 150/ 151 - Node out of scope
#> TARGET : reaction_2018659 151/ 151 - Node out of scope
#> Extracting non-random path p<0.1 for E2A/PBX1
#> There are 151 nodes in the neighborhoodTARGET : reaction_71850 1/ 151 - Found a path to reaction_71850
#> TARGET : reaction_71496 2/ 151 - Node out of scope
#> TARGET : reaction_2046222 3/ 151 - Node out of scope
#> TARGET : reaction_2162226 4/ 151 - Node out of scope
#> TARGET : reaction_70403 5/ 151 - Found a path to reaction_70403
#> TARGET : reaction_2162227 6/ 151 - Node out of scope
#> TARGET : reaction_2162225 7/ 151 - Node out of scope
#> TARGET : reaction_2090085 8/ 151 - Node out of scope
#> TARGET : reaction_3262512 9/ 151 - Node out of scope
#> TARGET : reaction_372448 10/ 151 - Node out of scope
#> TARGET : reaction_372449 11/ 151 - Found a path to reaction_372449
#> TARGET : reaction_71495 12/ 151 - Node out of scope
#> TARGET : reaction_2076508 13/ 151 - Node out of scope
#> TARGET : reaction_71515 14/ 151 - Node out of scope
#> TARGET : reaction_1638032 15/ 151 - Node out of scope
#> TARGET : reaction_70420 16/ 151 - Found a path to reaction_70420
#> TARGET : reaction_1678716 17/ 151 - Node out of scope
#> TARGET : reaction_1793217 18/ 151 - Node out of scope
#> TARGET : reaction_450092 19/ 151 - Found a path to reaction_450092
#> TARGET : reaction_450095 20/ 151 - Node out of scope
#> TARGET : reaction_1678708 21/ 151 - Node out of scope
#> TARGET : reaction_70427 22/ 151 - Node out of scope
#> TARGET : reaction_450088 23/ 151 - Found a path to reaction_450088
#> TARGET : reaction_3266566 24/ 151 - Node out of scope
#> TARGET : reaction_191101 25/ 151 - Node out of scope
#> TARGET : reaction_198508 26/ 151 - Node out of scope
#> TARGET : reaction_1678660 27/ 151 - Node out of scope
#> TARGET : reaction_1630306 28/ 151 - Node out of scope
#> TARGET : reaction_1630304 29/ 151 - Node out of scope
#> TARGET : reaction_70454 30/ 151 - Node out of scope
#> TARGET : reaction_70449 31/ 151 - Node out of scope
#> TARGET : reaction_2046265 32/ 151 - Found a path to reaction_2046265
#> TARGET : reaction_71825 33/ 151 - Node out of scope
#> TARGET : reaction_198513 34/ 151 - Found a path to reaction_198513
#> TARGET : reaction_71541 35/ 151 - Found a path to reaction_71541
#> TARGET : reaction_70467 36/ 151 - Found a path to reaction_70467
#> TARGET : reaction_1793182 37/ 151 - Node out of scope
#> TARGET : reaction_2022061 38/ 151 - Node out of scope
#> TARGET : reaction_70471 39/ 151 - Found a path to reaction_70471
#> TARGET : reaction_2046285 40/ 151 - Found a path to reaction_2046285
#> TARGET : reaction_70475 41/ 151 - Found a path to reaction_70475
#> TARGET : reaction_2022052 42/ 151 - Node out of scope
#> TARGET : reaction_70479 43/ 151 - Node out of scope
#> TARGET : reaction_2025724 44/ 151 - Found a path to reaction_2025724
#> TARGET : reaction_1678650 45/ 151 - Node out of scope
#> TARGET : reaction_70482 46/ 151 - Node out of scope
#> TARGET : reaction_741449 47/ 151 - Node out of scope
#> TARGET : reaction_70481 48/ 151 - Node out of scope
#> TARGET : reaction_2046298 49/ 151 - Found a path to reaction_2046298
#> TARGET : reaction_2090043 50/ 151 - Node out of scope
#> TARGET : reaction_372843 51/ 151 - Node out of scope
#> TARGET : reaction_2022851 52/ 151 - Found a path to reaction_2022851
#> TARGET : reaction_70486 53/ 151 - Found a path to reaction_70486
#> TARGET : reaction_1606789 54/ 151 - Node out of scope
#> TARGET : reaction_1678854 55/ 151 - Node out of scope
#> TARGET : reaction_2105001 56/ 151 - Node out of scope
#> TARGET : reaction_2160915 57/ 151 - Node out of scope
#> TARGET : reaction_177784 58/ 151 - Node out of scope
#> TARGET : reaction_2022860 59/ 151 - Node out of scope
#> TARGET : reaction_71445 60/ 151 - Node out of scope
#> TARGET : reaction_2090037 61/ 151 - Node out of scope
#> TARGET : reaction_70494 62/ 151 - Node out of scope
#> TARGET : reaction_2090038 63/ 151 - Node out of scope
#> TARGET : reaction_2022856 64/ 151 - Found a path to reaction_2022856
#> TARGET : reaction_70262 65/ 151 - Found a path to reaction_70262
#> TARGET : reaction_70501 66/ 151 - Node out of scope
#> TARGET : reaction_198440 67/ 151 - Node out of scope
#> TARGET : reaction_372819 68/ 151 - Node out of scope
#> TARGET : reaction_189102 69/ 151 - Node out of scope
#> TARGET : reaction_70241 70/ 151 - Node out of scope
#> TARGET : reaction_2022887 71/ 151 - Found a path to reaction_2022887
#> TARGET : reaction_198458 72/ 151 - Node out of scope
#> TARGET : reaction_2076371 73/ 151 - Node out of scope
#> TARGET : reaction_163750 74/ 151 - Node out of scope
#> TARGET : reaction_2065233 75/ 151 - Node out of scope
#> TARGET : reaction_2076611 76/ 151 - Node out of scope
#> TARGET : reaction_163751 77/ 151 - Found a path to reaction_163751
#> TARGET : reaction_2076383 78/ 151 - Node out of scope
#> TARGET : reaction_70290 79/ 151 - Node out of scope
#> TARGET : reaction_70286 80/ 151 - Node out of scope
#> TARGET : reaction_163764 81/ 151 - Found a path to reaction_163764
#> TARGET : reaction_199803 82/ 151 - Found a path to reaction_199803
#> TARGET : reaction_1971487 83/ 151 - Node out of scope
#> TARGET : reaction_163773 84/ 151 - Node out of scope
#> TARGET : reaction_70272 85/ 151 - Found a path to reaction_70272
#> TARGET : reaction_1971483 86/ 151 - Node out of scope
#> TARGET : reaction_70333 87/ 151 - Node out of scope
#> TARGET : reaction_2160884 88/ 151 - Node out of scope
#> TARGET : reaction_1971491 89/ 151 - Node out of scope
#> TARGET : reaction_71654 90/ 151 - Node out of scope
#> TARGET : reaction_1889981 91/ 151 - Node out of scope
#> TARGET : reaction_2160892 92/ 151 - Node out of scope
#> TARGET : reaction_71660 93/ 151 - Node out of scope
#> TARGET : reaction_1889978 94/ 151 - Node out of scope
#> TARGET : reaction_189053 95/ 151 - Node out of scope
#> TARGET : reaction_73580 96/ 151 - Node out of scope
#> TARGET : reaction_71670 97/ 151 - Found a path to reaction_71670
#> TARGET : reaction_2024100 98/ 151 - Node out of scope
#> TARGET : reaction_70581 99/ 151 - Node out of scope
#> TARGET : reaction_2076392 100/ 151 - Found a path to reaction_2076392
#> TARGET : reaction_2160874 101/ 151 - Node out of scope
#> TARGET : reaction_163741 102/ 151 - Node out of scope
#> TARGET : reaction_70361 103/ 151 - Node out of scope
#> TARGET : reaction_71324 104/ 151 - Found a path to reaction_71324
#> TARGET : reaction_453337 105/ 151 - Found a path to reaction_453337
#> TARGET : reaction_453338 106/ 151 - Node out of scope
#> TARGET : reaction_70355 107/ 151 - Found a path to reaction_70355
#> TARGET : reaction_70592 108/ 151 - Node out of scope
#> TARGET : reaction_453339 109/ 151 - Node out of scope
#> TARGET : reaction_2046175 110/ 151 - Node out of scope
#> TARGET : reaction_453342 111/ 151 - Node out of scope
#> TARGET : reaction_70596 112/ 151 - Node out of scope
#> TARGET : reaction_71574 113/ 151 - Node out of scope
#> TARGET : reaction_174392 114/ 151 - Node out of scope
#> TARGET : reaction_71303 115/ 151 - Found a path to reaction_71303
#> TARGET : reaction_70349 116/ 151 - Found a path to reaction_70349
#> TARGET : reaction_191108 117/ 151 - Node out of scope
#> TARGET : reaction_71296 118/ 151 - Node out of scope
#> TARGET : reaction_71299 119/ 151 - Node out of scope
#> TARGET : reaction_174389 120/ 151 - Node out of scope
#> TARGET : reaction_71802 121/ 151 - Found a path to reaction_71802
#> TARGET : reaction_191116 122/ 151 - Node out of scope
#> TARGET : reaction_70613 123/ 151 - Found a path to reaction_70613
#> TARGET : reaction_71306 124/ 151 - Node out of scope
#> TARGET : reaction_70342 125/ 151 - Node out of scope
#> TARGET : reaction_71588 126/ 151 - Found a path to reaction_71588
#> TARGET : reaction_71590 127/ 151 - Node out of scope
#> TARGET : reaction_2142859 128/ 151 - Node out of scope
#> TARGET : reaction_71593 129/ 151 - Node out of scope
#> TARGET : reaction_71335 130/ 151 - Found a path to reaction_71335
#> TARGET : reaction_453346 131/ 151 - Node out of scope
#> TARGET : reaction_71334 132/ 151 - Node out of scope
#> TARGET : reaction_70377 133/ 151 - Found a path to reaction_70377
#> TARGET : reaction_2046180 134/ 151 - Node out of scope
#> TARGET : reaction_71602 135/ 151 - Node out of scope
#> TARGET : reaction_1678742 136/ 151 - Node out of scope
#> TARGET : reaction_70369 137/ 151 - Node out of scope
#> TARGET : reaction_453358 138/ 151 - Node out of scope
#> TARGET : reaction_189222 139/ 151 - Node out of scope
#> TARGET : reaction_453356 140/ 151 - Node out of scope
#> TARGET : reaction_1638053 141/ 151 - Node out of scope
#> TARGET : reaction_2022063 142/ 151 - Node out of scope
#> TARGET : reaction_2090079 143/ 151 - Node out of scope
#> TARGET : reaction_1793186 144/ 151 - Node out of scope
#> TARGET : reaction_2076419 145/ 151 - Node out of scope
#> TARGET : reaction_1793207 146/ 151 - Node out of scope
#> TARGET : reaction_1889955 147/ 151 - Node out of scope
#> TARGET : reaction_111215 148/ 151 - Node out of scope
#> TARGET : reaction_71610 149/ 151 - Node out of scope
#> TARGET : reaction_2018682 150/ 151 - Node out of scope
#> TARGET : reaction_2018659 151/ 151 - Node out of scope
#> Extracting non-random path p<0.1 for NEG
#> There are 151 nodes in the neighborhoodTARGET : reaction_71850 1/ 151 - Node out of scope
#> TARGET : reaction_71496 2/ 151 - Node out of scope
#> TARGET : reaction_2046222 3/ 151 - Node out of scope
#> TARGET : reaction_2162226 4/ 151 - Node out of scope
#> TARGET : reaction_70403 5/ 151 - Found a path to reaction_70403
#> TARGET : reaction_2162227 6/ 151 - Node out of scope
#> TARGET : reaction_2162225 7/ 151 - Node out of scope
#> TARGET : reaction_2090085 8/ 151 - Node out of scope
#> TARGET : reaction_3262512 9/ 151 - Node out of scope
#> TARGET : reaction_372448 10/ 151 - Node out of scope
#> TARGET : reaction_372449 11/ 151 - Node out of scope
#> TARGET : reaction_71495 12/ 151 - Node out of scope
#> TARGET : reaction_2076508 13/ 151 - Node out of scope
#> TARGET : reaction_71515 14/ 151 - Node out of scope
#> TARGET : reaction_1638032 15/ 151 - Node out of scope
#> TARGET : reaction_70420 16/ 151 - Node out of scope
#> TARGET : reaction_1678716 17/ 151 - Node out of scope
#> TARGET : reaction_1793217 18/ 151 - Node out of scope
#> TARGET : reaction_450092 19/ 151 - Node out of scope
#> TARGET : reaction_450095 20/ 151 - Node out of scope
#> TARGET : reaction_1678708 21/ 151 - Node out of scope
#> TARGET : reaction_70427 22/ 151 - Node out of scope
#> TARGET : reaction_450088 23/ 151 - Node out of scope
#> TARGET : reaction_3266566 24/ 151 - Node out of scope
#> TARGET : reaction_191101 25/ 151 - Node out of scope
#> TARGET : reaction_198508 26/ 151 - Node out of scope
#> TARGET : reaction_1678660 27/ 151 - Node out of scope
#> TARGET : reaction_1630306 28/ 151 - Node out of scope
#> TARGET : reaction_1630304 29/ 151 - Node out of scope
#> TARGET : reaction_70454 30/ 151 - Node out of scope
#> TARGET : reaction_70449 31/ 151 - Node out of scope
#> TARGET : reaction_2046265 32/ 151 - Node out of scope
#> TARGET : reaction_71825 33/ 151 - Node out of scope
#> TARGET : reaction_198513 34/ 151 - Node out of scope
#> TARGET : reaction_71541 35/ 151 - Node out of scope
#> TARGET : reaction_70467 36/ 151 - Node out of scope
#> TARGET : reaction_1793182 37/ 151 - Node out of scope
#> TARGET : reaction_2022061 38/ 151 - Node out of scope
#> TARGET : reaction_70471 39/ 151 - Node out of scope
#> TARGET : reaction_2046285 40/ 151 - Node out of scope
#> TARGET : reaction_70475 41/ 151 - Node out of scope
#> TARGET : reaction_2022052 42/ 151 - Node out of scope
#> TARGET : reaction_70479 43/ 151 - Node out of scope
#> TARGET : reaction_2025724 44/ 151 - Node out of scope
#> TARGET : reaction_1678650 45/ 151 - Node out of scope
#> TARGET : reaction_70482 46/ 151 - Node out of scope
#> TARGET : reaction_741449 47/ 151 - Node out of scope
#> TARGET : reaction_70481 48/ 151 - Node out of scope
#> TARGET : reaction_2046298 49/ 151 - Node out of scope
#> TARGET : reaction_2090043 50/ 151 - Node out of scope
#> TARGET : reaction_372843 51/ 151 - Node out of scope
#> TARGET : reaction_2022851 52/ 151 - Node out of scope
#> TARGET : reaction_70486 53/ 151 - Node out of scope
#> TARGET : reaction_1606789 54/ 151 - Node out of scope
#> TARGET : reaction_1678854 55/ 151 - Node out of scope
#> TARGET : reaction_2105001 56/ 151 - Node out of scope
#> TARGET : reaction_2160915 57/ 151 - Node out of scope
#> TARGET : reaction_177784 58/ 151 - Node out of scope
#> TARGET : reaction_2022860 59/ 151 - Node out of scope
#> TARGET : reaction_71445 60/ 151 - Node out of scope
#> TARGET : reaction_2090037 61/ 151 - Node out of scope
#> TARGET : reaction_70494 62/ 151 - Node out of scope
#> TARGET : reaction_2090038 63/ 151 - Node out of scope
#> TARGET : reaction_2022856 64/ 151 - Node out of scope
#> TARGET : reaction_70262 65/ 151 - Node out of scope
#> TARGET : reaction_70501 66/ 151 - Node out of scope
#> TARGET : reaction_198440 67/ 151 - Node out of scope
#> TARGET : reaction_372819 68/ 151 - Node out of scope
#> TARGET : reaction_189102 69/ 151 - Node out of scope
#> TARGET : reaction_70241 70/ 151 - Node out of scope
#> TARGET : reaction_2022887 71/ 151 - Node out of scope
#> TARGET : reaction_198458 72/ 151 - Node out of scope
#> TARGET : reaction_2076371 73/ 151 - Node out of scope
#> TARGET : reaction_163750 74/ 151 - Node out of scope
#> TARGET : reaction_2065233 75/ 151 - Node out of scope
#> TARGET : reaction_2076611 76/ 151 - Node out of scope
#> TARGET : reaction_163751 77/ 151 - Node out of scope
#> TARGET : reaction_2076383 78/ 151 - Node out of scope
#> TARGET : reaction_70290 79/ 151 - Node out of scope
#> TARGET : reaction_70286 80/ 151 - Node out of scope
#> TARGET : reaction_163764 81/ 151 - Node out of scope
#> TARGET : reaction_199803 82/ 151 - Node out of scope
#> TARGET : reaction_1971487 83/ 151 - Node out of scope
#> TARGET : reaction_163773 84/ 151 - Node out of scope
#> TARGET : reaction_70272 85/ 151 - Node out of scope
#> TARGET : reaction_1971483 86/ 151 - Node out of scope
#> TARGET : reaction_70333 87/ 151 - Node out of scope
#> TARGET : reaction_2160884 88/ 151 - Node out of scope
#> TARGET : reaction_1971491 89/ 151 - Node out of scope
#> TARGET : reaction_71654 90/ 151 - Node out of scope
#> TARGET : reaction_1889981 91/ 151 - Node out of scope
#> TARGET : reaction_2160892 92/ 151 - Node out of scope
#> TARGET : reaction_71660 93/ 151 - Node out of scope
#> TARGET : reaction_1889978 94/ 151 - Node out of scope
#> TARGET : reaction_189053 95/ 151 - Node out of scope
#> TARGET : reaction_73580 96/ 151 - Node out of scope
#> TARGET : reaction_71670 97/ 151 - Node out of scope
#> TARGET : reaction_2024100 98/ 151 - Node out of scope
#> TARGET : reaction_70581 99/ 151 - Node out of scope
#> TARGET : reaction_2076392 100/ 151 - Node out of scope
#> TARGET : reaction_2160874 101/ 151 - Node out of scope
#> TARGET : reaction_163741 102/ 151 - Node out of scope
#> TARGET : reaction_70361 103/ 151 - Node out of scope
#> TARGET : reaction_71324 104/ 151 - Node out of scope
#> TARGET : reaction_453337 105/ 151 - Node out of scope
#> TARGET : reaction_453338 106/ 151 - Node out of scope
#> TARGET : reaction_70355 107/ 151 - Node out of scope
#> TARGET : reaction_70592 108/ 151 - Node out of scope
#> TARGET : reaction_453339 109/ 151 - Node out of scope
#> TARGET : reaction_2046175 110/ 151 - Node out of scope
#> TARGET : reaction_453342 111/ 151 - Node out of scope
#> TARGET : reaction_70596 112/ 151 - Node out of scope
#> TARGET : reaction_71574 113/ 151 - Node out of scope
#> TARGET : reaction_174392 114/ 151 - Node out of scope
#> TARGET : reaction_71303 115/ 151 - Node out of scope
#> TARGET : reaction_70349 116/ 151 - Node out of scope
#> TARGET : reaction_191108 117/ 151 - Node out of scope
#> TARGET : reaction_71296 118/ 151 - Node out of scope
#> TARGET : reaction_71299 119/ 151 - Node out of scope
#> TARGET : reaction_174389 120/ 151 - Node out of scope
#> TARGET : reaction_71802 121/ 151 - Node out of scope
#> TARGET : reaction_191116 122/ 151 - Node out of scope
#> TARGET : reaction_70613 123/ 151 - Node out of scope
#> TARGET : reaction_71306 124/ 151 - Node out of scope
#> TARGET : reaction_70342 125/ 151 - Node out of scope
#> TARGET : reaction_71588 126/ 151 - Node out of scope
#> TARGET : reaction_71590 127/ 151 - Node out of scope
#> TARGET : reaction_2142859 128/ 151 - Node out of scope
#> TARGET : reaction_71593 129/ 151 - Node out of scope
#> TARGET : reaction_71335 130/ 151 - Node out of scope
#> TARGET : reaction_453346 131/ 151 - Node out of scope
#> TARGET : reaction_71334 132/ 151 - Node out of scope
#> TARGET : reaction_70377 133/ 151 - Node out of scope
#> TARGET : reaction_2046180 134/ 151 - Node out of scope
#> TARGET : reaction_71602 135/ 151 - Node out of scope
#> TARGET : reaction_1678742 136/ 151 - Node out of scope
#> TARGET : reaction_70369 137/ 151 - Node out of scope
#> TARGET : reaction_453358 138/ 151 - Node out of scope
#> TARGET : reaction_189222 139/ 151 - Node out of scope
#> TARGET : reaction_453356 140/ 151 - Node out of scope
#> TARGET : reaction_1638053 141/ 151 - Node out of scope
#> TARGET : reaction_2022063 142/ 151 - Node out of scope
#> TARGET : reaction_2090079 143/ 151 - Node out of scope
#> TARGET : reaction_1793186 144/ 151 - Node out of scope
#> TARGET : reaction_2076419 145/ 151 - Node out of scope
#> TARGET : reaction_1793207 146/ 151 - Node out of scope
#> TARGET : reaction_1889955 147/ 151 - Node out of scope
#> TARGET : reaction_111215 148/ 151 - Node out of scope
#> TARGET : reaction_71610 149/ 151 - Node out of scope
#> TARGET : reaction_2018682 150/ 151 - Node out of scope
#> TARGET : reaction_2018659 151/ 151 - Node out of scope